+++ src/gdal_utils.cpp
@@ -17,6 +17,7 @@ void Utils::Initialize(Local<Object> target) {
   Nan__SetAsyncableMethod(target, "info", info);
   Nan__SetAsyncableMethod(target, "translate", translate);
   Nan__SetAsyncableMethod(target, "vectorTranslate", vectorTranslate);
+  Nan__SetAsyncableMethod(target, "warp", warp);
 }
 
 /**
@@ -32,7 +33,7 @@ void Utils::Initialize(Local<Object> target) {
  * @static
  * @param {string} destination destination filename
  * @param {gdal.Dataset} source source dataset
- * @param {string[]} array of CLI options for gdal_translate
+ * @param {string[]} [options] array of CLI options for gdal_translate
  * @return {gdal.Dataset}
  */
 
@@ -50,7 +51,7 @@ void Utils::Initialize(Local<Object> target) {
  * @static
  * @param {string} destination destination filename
  * @param {gdal.Dataset} source source dataset
- * @param {string[]} array of CLI options for gdal_translate
+ * @param {string[]} [options] array of CLI options for gdal_translate
  * @param {callback<void>} [callback=undefined] {{{cb}}}
  * @return {Promise<gdal.Dataset>}
  */
@@ -66,11 +67,11 @@ GDAL_ASYNCABLE_DEFINE(Utils::translate) {
   GDAL_RAW_CHECK(GDALDataset *, ds, raw);
 
   Local<Array> options;
-  NODE_ARG_ARRAY(2, "options", options);
-
-  for (unsigned i = 0; i < options->Length(); ++i) {
-    aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
-  }
+  NODE_ARG_ARRAY_OPT(2, "options", options);
+  if (!options.IsEmpty())
+    for (unsigned i = 0; i < options->Length(); ++i) {
+      aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
+    }
 
   GDALAsyncableJob<GDALDataset *> job(ds->uid);
   job.main = [raw, dst, aosOptions](const GDALExecutionProgress &) {
@@ -100,7 +101,7 @@ GDAL_ASYNCABLE_DEFINE(Utils::translate) {
  * @static
  * @param {string|gdal.Dataset} destination destination
  * @param {gdal.Dataset} source source dataset
- * @param {string[]} array of CLI options for ogr2ogr
+ * @param {string[]} [options] array of CLI options for ogr2ogr
  * @return {gdal.Dataset}
  */
 
@@ -118,7 +119,7 @@ GDAL_ASYNCABLE_DEFINE(Utils::translate) {
  * @static
  * @param {string|gdal.Dataset} destination destination
  * @param {gdal.Dataset} source source dataset
- * @param {string[]} array of CLI options for ogr2ogr
+ * @param {string[]} [options] array of CLI options for ogr2ogr
  * @param {callback<void>} [callback=undefined] {{{cb}}}
  * @return {Promise<gdal.Dataset>}
  */
@@ -150,11 +151,11 @@ GDAL_ASYNCABLE_DEFINE(Utils::vectorTranslate) {
   GDAL_RAW_CHECK(GDALDataset *, ds, src_raw);
 
   Local<Array> options;
-  NODE_ARG_ARRAY(2, "options", options);
-
-  for (unsigned i = 0; i < options->Length(); ++i) {
-    aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
-  }
+  NODE_ARG_ARRAY_OPT(2, "options", options);
+  if (!options.IsEmpty())
+    for (unsigned i = 0; i < options->Length(); ++i) {
+      aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
+    }
 
   std::vector<long> uids = {ds->uid};
   if (dst_ds != nullptr) uids.push_back(dst_ds->uid);
@@ -162,7 +163,6 @@ GDAL_ASYNCABLE_DEFINE(Utils::vectorTranslate) {
 
   job.main = [src_raw, dst_filename, dst_raw, aosOptions](const GDALExecutionProgress &) {
     CPLErrorReset();
-    auto b = aosOptions;
     auto psOptions = GDALVectorTranslateOptionsNew(aosOptions->List(), nullptr);
 
     auto srcH = GDALDatasetToHandle(src_raw);
@@ -190,7 +190,7 @@ GDAL_ASYNCABLE_DEFINE(Utils::vectorTranslate) {
  * @for gdal
  * @static
  * @param {gdal.Dataset} dataset
- * @param {string[]} array of CLI options for gdalinfo
+ * @param {string[]} [options] array of CLI options for gdalinfo
  * @return {string}
  */
 
@@ -207,7 +207,7 @@ GDAL_ASYNCABLE_DEFINE(Utils::vectorTranslate) {
  * @for gdal
  * @static
  * @param {gdal.Dataset} dataset
- * @param {string[]} array of CLI options for gdalinfo
+ * @param {string[]} [options] array of CLI options for gdalinfo
  * @param {callback<void>} [callback=undefined] {{{cb}}}
  * @return {Promise<string>}
  */
@@ -220,16 +220,15 @@ GDAL_ASYNCABLE_DEFINE(Utils::info) {
   GDAL_RAW_CHECK(GDALDataset *, ds, raw);
 
   Local<Array> options;
-  NODE_ARG_ARRAY(1, "options", options);
-
-  for (unsigned i = 0; i < options->Length(); ++i) {
-    aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
-  }
+  NODE_ARG_ARRAY_OPT(1, "options", options);
+  if (!options.IsEmpty())
+    for (unsigned i = 0; i < options->Length(); ++i) {
+      aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
+    }
 
   GDALAsyncableJob<std::string> job(ds->uid);
   job.main = [raw, aosOptions](const GDALExecutionProgress &) {
     CPLErrorReset();
-    auto b = aosOptions;
     auto psOptions = GDALInfoOptionsNew(aosOptions->List(), nullptr);
     char *r = GDALInfo(GDALDatasetToHandle(raw), psOptions);
     GDALInfoOptionsFree(psOptions);
@@ -241,6 +240,107 @@ GDAL_ASYNCABLE_DEFINE(Utils::info) {
   job.rval = [](std::string s, GetFromPersistentFunc) { return SafeString::New(s.c_str()); };
 
   job.run(info, async, 2);
+}
+
+/**
+ * Library version of gdalwarp.
+ *
+ * @example
+ * const ds = gdal.open('input.tif')
+ * const output = gdal.warp('/vsimem/temp.tif')
+ *
+ * @throws Error
+ * @method warp
+ * @for gdal
+ * @static
+ * @param {string} [dst_path] destination path, null for an in-memory operation
+ * @param {gdal.Dataset} [dst_ds] destination dataset, null for a new dataset
+ * @param {gdal.Dataset[]} src_ds array of source datasets
+ * @param {string[]} [options] array of CLI options for gdalwarp
+ * @return {gdal.Dataset}
+ */
+
+/**
+ * Library version of gdalinfo.
+ * {{{async}}}
+ *
+ * @example
+ * const ds = gdal.open('input.tif')
+ * const output = gdal.warp('/vsimem/temp.tif')
+ * @throws Error
+ *
+ * @method warpAsync
+ * @for gdal
+ * @static
+ * @param {string} [dst_path] destination path, null for an in-memory operation
+ * @param {gdal.Dataset} [dst_ds] destination dataset, null for a new dataset
+ * @param {gdal.Dataset[]} src_ds array of source datasets
+ * @param {string[]} [options] array of CLI options for gdalwarp
+ * @param {callback<void>} [callback=undefined] {{{cb}}}
+ * @return {Promise<gdal.Dataset>}
+ */
+GDAL_ASYNCABLE_DEFINE(Utils::warp) {
+  auto aosOptions = std::make_shared<CPLStringList>();
+
+  std::string dst_path("");
+  Local<Object> dst_ds;
+
+  NODE_ARG_OPT_STR(0, "dst_path", dst_path);
+
+  NODE_ARG_OBJECT_OPT(1, "dst_ds", dst_ds);
+  GDALDatasetH gdal_dst_ds = nullptr;
+  if (!dst_ds.IsEmpty()) {
+    NODE_UNWRAP_CHECK(Dataset, dst_ds, ds);
+    GDAL_RAW_CHECK(GDALDataset *, ds, raw);
+    gdal_dst_ds = GDALDatasetToHandle(raw);
+  }
+
+  if (dst_path.length() == 0 && gdal_dst_ds == nullptr) {
+    Nan::ThrowError("Either \"dst_path\" or \"dst_ds\" must be given");
+    return;
+  }
+
+  Local<Array> src_ds;
+  NODE_ARG_ARRAY(2, "src_ds", src_ds);
+  if (src_ds->Length() < 1) {
+    Nan::ThrowError("\"src_ds\" must contain at least one element");
+    return;
+  }
+  auto gdal_src_ds = std::shared_ptr<GDALDatasetH>(new GDALDatasetH[src_ds->Length()], array_deleter<GDALDatasetH>());
+  std::vector<long> uids;
+  for (unsigned i = 0; i < src_ds->Length(); ++i) {
+    NODE_UNWRAP_CHECK(Dataset, Nan::Get(src_ds, i).ToLocalChecked().As<Object>(), ds);
+    GDAL_RAW_CHECK(GDALDataset *, ds, raw);
+    gdal_src_ds.get()[i] = GDALDatasetToHandle(raw);
+    uids.push_back(ds->uid);
+  }
+
+  Local<Array> options;
+  NODE_ARG_ARRAY_OPT(3, "options", options);
+  if (!options.IsEmpty())
+    for (unsigned i = 0; i < options->Length(); ++i) {
+      aosOptions->AddString(*Nan::Utf8String(Nan::Get(options, i).ToLocalChecked()));
+    }
+
+  GDALAsyncableJob<GDALDataset *> job(uids);
+  int src_count = src_ds->Length();
+  job.main = [dst_path, gdal_dst_ds, src_count, gdal_src_ds, aosOptions](const GDALExecutionProgress &) {
+    CPLErrorReset();
+    auto psOptions = GDALWarpAppOptionsNew(aosOptions->List(), nullptr);
+    GDALDatasetH r = GDALWarp(
+      dst_path.length() > 0 ? dst_path.c_str() : nullptr,
+      gdal_dst_ds,
+      src_count,
+      gdal_src_ds.get(),
+      psOptions,
+      nullptr);
+    GDALWarpAppOptionsFree(psOptions);
+    if (r == nullptr) throw CPLGetLastErrorMsg();
+    return GDALDatasetFromHandle(r);
+  };
+  job.rval = [](GDALDataset *ds, GetFromPersistentFunc) { return Dataset::New(ds); };
+
+  job.run(info, async, 4);
 }
 
 } // namespace node_gdal